Adapt the CHANGELOG to the Common Changelog format (#59)
So far the `CHANGELOG.md` file followed "Keep a Changelog", which is
somewhat underspecified; this led to inconsistent-looking changelog
entries. "Common Changelog" is a stricter variant which makes it
explicit how the entries should be written. In this PR, I'm switching to
the latter, and adjusting existing changelog entries to match that.