Expand README and relax protobuf version requirement (#62)
This PR expands on the Tensorflow troubleshooting section in
`README.md`, taking into account how installing the newest versions on
Ubuntu 22.04 requires extra care (fixes #61). On top of this, I also
relaxed the pin on `protobuf` version in `setup.py`; I'm not sure why
the lower bound was introduced, but some of the `tensorflow` versions
actually conflict with it.