Fix hydrogen handling in scaffolds with explicit attachment points (#70)
When a scaffold is provided during decoding, we assume the model is free
to attach further atoms to any atom in the scaffold, unless it contains
explicit attachment points, in which case we only allow attaching to
those atoms. However, the handling of these attachment points had a bug
regarding explicit hydrogens, which would cause the scaffold molecule to
be malformed and decoding to fail. This PR fixes this issue by carefully
updating the hydrogen count where appropriate.